Music regex to xml parser Add new feature on Music : Add get basic information from music currently streamed by VLC server A minor bug found when title != filename, no solution for now. Maybe force to write title = filename or filename = title depend of the relevance and for all song ~~
Music feature for first phase is complete for me. To manage metadata is for later after third deadline...
I can test with Manageplaylist now.

+++ Gestion/Music.py
@@ -8,11 +8,22 @@ import subprocess
 import re
 from Gestion import Ctes
 from Gestion.Enum import *
+import xml.etree.cElementTree as ET
+from pathlib import Path
 
 path = ''
 pwd_vlc = ''
-chemin = ''
+chemin = '/run/media/vinsento/455B532D7757A5FD/Project/MultiRoom/status.xml'
 user_vlc = ''
+
+class MusicMetadata:
+
+    def __init__(self):
+        self.artist = ""
+        self.album = ""
+
+    def updateMetadata(self, pathCurrentSong):
+        return ReturnCode.ErrNotImplemented
 
 class Music():
     '''
@@ -20,49 +31,73 @@ class Music():
     Works only on Linus OS
     '''
 
-    @staticmethod
-    def name():
+    def __init__(self, portCtrl):
+        self.currentMusic = ""
+        self.path = ""
+        self.metadata = MusicMetadata()
+        self.portCtrl = str(portCtrl)
+        self.count = 0
+        self.cmd = 'wget --http-user=' + Ctes.user_vlc + ' --http-password=' + Ctes.pwd_vlc + ' ' + ' ' + Ctes.local_ip + ':' + self.portCtrl + '/requests/status.xml'
+
+    def updateInfo(self):
+        self.currentMusic = self.getNameMusic()
+        self.path = self.getPath()
+
+
+    def getNameMusic(self):
         """
         Retrieve the name of the current song play in VLC
         :return:
         """
         try:
-            os.system('wget http://127.0.0.1:8080/requests/status.xml --http-user=' + Ctes.user_vlc + ' --http-password=' + Ctes.pwd_vlc)
-            file = open(chemin, 'r')
-            a = file.read()
-            # print(a)  a=re.search(r'a','fdfdffdabavfgfgf' test)
+            if os.path.isfile(chemin):
+                os.system('rm status.xml')
+            os.system(self.cmd)
+            print(self.cmd)
 
-            pos_rep1 = re.search(r"<info name=>", a)
-            debut = pos_rep1.span()[1]
-            pos_rep2 = re.search(r"mp3", a)
-            fin = pos_rep2.span()[1]
-            os.system('rm status.xml')
-            return a[debut:fin]
+
+            if not os.path.isfile(chemin):
+                return ReturnCode.ErrNotImplemented
+
+            tree = ET.parse(chemin)
+            root = tree.getroot()
+            nodeInformation = root.find("information")
+            for nodeInfo in nodeInformation[0]:
+                if nodeInfo.attrib["name"] == "filename":
+                    self.currentMusic = nodeInfo.text
+                    return nodeInfo.text
+
+
         except:
             print('File Not Found or VLC is not running')
-        return
+        return "toto"
 
-    @staticmethod
-    def delMusic():
+
+    def delMusic(self):
+        """
+        Need to delete in all playlist also....
+        :return:
+        """
         return ReturnCode.ErrNotImplemented
 
-    @staticmethod
-    def path():
+    def getPath(self):
         """
         Retrieve the path of the current song play in VLC from database contained all path of all files in the system
         :return:
         """
-        nom = Music.name()
+
         try:
-            p = subprocess.check_output(["locate", nom, "--database", "external.red.db"])
+            p = subprocess.check_output(["locate", self.currentMusic, "--database", "external.red.db"])
             p = p.splitlines()
             path = ''
             # v.decode() car le fichier est en binaire va savoir pourquoi...Flem de lire la doc
             for v in p:
-                if nom in v.decode():
+                currentPath = v.decode().split("/")
+                if self.currentMusic in currentPath[-1]:
                     # Cela serai bien de trouver des doublons
                     # Avec un test regex pour le chemin
                     path = v.decode()
+                    break
             # Raise une erreur cela serai mieux
             if path is not '':
                 return path
@@ -70,7 +105,11 @@ class Music():
         except:
             print('Exception : File Not Found')
             Music.maj_db()
-            Music.path()
+            self.count = self.count + 1
+            if self.count > 5:
+                print(self.currentMusic)
+                return "Error"
+            self.getPath()
 
     @staticmethod
     def maj_db():
@@ -78,4 +117,10 @@ class Music():
         Update the database witch record all files with their path in a file in binary
         :return:
         """
-        os.system('sudo updatedb -o external.red.db -U /mnt/NAS/')
+        os.system('sudo updatedb -o external.red.db -U /run/media/vinsento/455B532D7757A5FD/Project/Test/')
+
+
+
+    def modifyMetaData(self):
+        return ReturnCode.ErrNotImplemented
+
